extracts db methods into another module
